Algorithm/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
ortools==9.8.3296
python-multipart==0.0.6
